Turn testing inside out
Now that we have standard-test-roles in Fedora, we can use that to
create virtual machines for us.

The role repositories don't need to include `run-tests.sh` anymore, and
as such don't depend on a random container. Instead, integration tests
can be run like this:

    export TEST_SUBJECTS=/path/to/image.qcow2
    ansible-playbook -i /usr/share/ansible/inventory/standard-inventory-qcow2 test/test_*.yml

We still need some way to run the tests on open pull requests. We used
semaphore for this before, but that has shortcomings. For one, each
repository needs to copy the test script and cannot share already
downloaded images with other repositories. Also, we cannot test against
private images (such as RHEL) on semaphore.

Thus, this commit changes most of the contents of this repository to
turn it into a test runner that can be run on a private OpenShift
instance. It polls GitHub for open pull requests, runs the tests as
above for multiple images, copies the results to an ssh server, and sets
GitHub's commit status.