Allow to specify arbitrary Ansible playbook in config.json.
This generalizes the images[i].setup command in config.json.

The playbook shall be given as a JSON array under the images[i].setup key in config.json. This is possible because Ansible playbooks are YAML and can be represented as JSON.

images[i].setup can be also a single command string for backward compatibility and simplicity.