Use Ansible for deployment; use BuildConfig for building
Added a playbook for deploying the CI environment to an
OpenShift cluster.

```
$ ansible-playbook -e test_harness_secrets_dir=$SECRETS_PATH \
    -e test_harness_config_dir=$CONFIG_PATH \
    -e test_harness_scc=$SCC_NAME \
    ansible/openshift-playbook.yml
```

Parameters:
* `test_harness_secrets_dir` - Required - See `SECRETS_PATH`
* `test_harness_config_dir` - Required - See `CONFIG_PATH`
* `test_harness_scc` - Required - See `SCC_NAME`
* `test_harness_namespace` - Default `lsr-test-harness`
* `test_harness_sa` - Default `system:serviceaccount:{{ test_harness_namespace }}:tester`
* `test_harness_need_node_selector` - Default `true`
* `test_harness_run_as_root` - Default `true`
* `test_harness_node_selector` - Default `{"system-roles-ci": "true"}`
* `test_harness_use_staging` - Default `true`
* `test_harness_use_production` - Default `false`

By default, this will deploy and/or update the staging environment
since`test_harness_use_staging: true`.  It will also change the
DeploymentConfig to use the nodeSelector above, and will configure the
DeploymentConfig to run as root.  If you want to deploy to production, you must
explicitly set `test_harness_use_production` to `true`.

I also added BuildConfigs for the staging and production images so
that we do not have a dependency on `docker.io` or other external
image registry.  I updated the `image` and `imagestream` definitions
to reflect this.