Add support for using ansible 2.7/python 2/centos7 as controller
This adds a container/pod for running tests using ansible 2.7 and
python 2 on centos7 as a controller node, in addition to the current
one which uses ansible 2.9/python3/Fedora 32.

The container is run in a new pod/deployment called
linux-system-roles-centos7 (and -staging).  This runs alongside
the current one.  The new one uses the same configuration as
the old one - that is, the linux-system-roles-centos7 deployment
pods will use configmap "config", and
linux-system-roles-centos7-staging will use "config-staging".

The config can be customized by the use of environment variables
in the deployment.  All of the run-tests command line arguments
have corresponding env. vars.  In addition, you can use env. vars.
to change the test variant and the list of linux-system-roles
repositories which will be tested.  See README.md for details.

The status in the PR now adds the ansible version like this:

centos-7/ansible-2.7
or
fedora-31/ansible-2.9 (staging)

This will be used for *all* statuses going forward.  That is,
there will be no more `linux-system-roles-test/fedora-31` - all
statuses, even for the current ansible, will include the ansible
version string.  The text in parentheses is the variant.  This
is controlled by `config["variant"]` in the configmap config.json,
or by using the `TEST_HARNESS_VARIANT` env. var.  If there
is no text in parenthesis, then this test is a production stage
test.  For staging, there will be "(staging)".  We might use
variant in the future for different test tools e.g. "(citool)"
or "(openstack)" or whatever, to distinguish tests running in
different environments.

We will need to change the configmaps used by CI to change
"name" to "variant".  For legacy support, if "name" is
"linux-system-roles-test", it is assumed the variant is
production so no "(name)" will be used.
If "name" is "linux-system-roles-test-staging" then
"(staging)" will be used.

There is now support for specifying that a platform requires
a minimum version of ansible in order to run.  There is a new
key 'min_ansible_version' available for each image.  For example,
since ansible 2.7 does not support managing a Fedora 32 host,
we can add `"min_ansible_version":"2.8"` to the Fedora 32 image
specification in the config, so that we will skip testing of
Fedora 32 when using ansible 2.7.

add cmdline args for `--variant` and `--repositories` so that
all settings have both a cmdline arg and a corresponding
environment variable.

Establish and document the precedence of settings: cmdline args,
env. vars., config.json, and defaults.