disable centos7/ansible 2.7 testing for now
LSR CI testing has a problem with hitting the github api ratelimit.
This is due to the design of the run-tests script.
Adding another deployment for ansible 2.7 testing just makes that
problem worse.  It seems that having another separate deployment
for ansible 2.7 is not the right approach.  We are in the process
of evaluating CI options.  If we stick with the current approach,
one alternative would be for the current deployment to launch
ansible-playbook for multiple versions of ansible.

However, this code may be useful, so for now not deleting it completely,
but leaving it in as disabled.