Ability to do a dry run with test_harness_dry_run=true
When testing, it is useful to be able to generate the OpenShift
files instead of applying the changes directly to the cluster.
When running ansible-playbook with e.g.
`-e test_harness_dry_run=true` you can dump the YAML files that
can be applied to the OpenShift cluster after review using e.g.
`oc apply -f <file>`