Derive image URL from compose URL
Can be used to support images whose URL is changing dynamically,
like latest Fedora cloud images.

Introduces a new option "compose" in the image dictionary,
which can be used instead of "source" and contains the compose URL.

Optionally it can be qualified by "variant" when the compose contains images of
multiple variants and specifying only compose would be ambiguous.