test-harness - Fixing a CI test failure
In the collection converted tests, the tempdirs to store converted
role and CI test files were not cleaned up every test cycle (per pr),
but only at the end. It left the extra files from the previous run
and when the leftover file is a test script, it caused a test failure.

With this fix, the tempdirs are generated and cleaned up for each pr.