have to use dumb-init to reap orphaned procs from std-inv-qcow2
because standard-inventory-qcow2 starts child processes and
orphans them, there is no way to call `wait` to reap them, and
the process table fills up with zombie/defunct processes.  Use
dumb-init which calls `wait`.

There is probably some way to use systemd for this . . . but it isn't
nearly as simple as calling `dumb-init -- some_process`